+++ Changelog.md
@@ -14,6 +14,8 @@
 - Improved VT backend performance (#342).
 - Improved text selection behaviour.
 - Improved detection of runtime DPI changes for KDE (Plasma) desktop environment.
+- Adds initial Gamepad input mapping support.
+- Adds `contour gamepad list` command to list all known available gamepads and their device identifier.
 - Adds preliminary implementation of `DA3` VT sequence.
 - Adds new configuration option `mouse_block_selection_modifier`.
 - Adds new configuration option `profiles.*.copy_last_mark_range_offset` (default `0`) to adjust where to start looking upwards for the `CopyPreviousMarkRange` action. This is useful for multi-line prompts.